Docs: rename and array folders and files of project as what I studied
I tried to adjust a convention of react that I searched. I will have to check if it is a right way.
