+++ src/components/Blog/BlogList.js
@@ -1,10 +1,7 @@
 /* eslint-disable no-unused-vars */
 import React, { useState } from 'react';
-import BlogPost from './BlogList/BlogPost';
 
-/** {component} List @경로 App/BlogList */
+/** Component */
 export default function BlogList(list) {
-  // state
-
   return <div className='blogList'></div>;
 }
