+++ src/components/Home.js
@@ -1,9 +1,8 @@
 /* eslint-disable no-unused-vars */
 import React, { useState } from 'react';
 
-/** {component} Home @경로 App/Body/Home */
+/** Component */
 export default function Home() {
-  // in div.body of App.js
   return (
     <div className='home'>
       <div className='viewer'>
