+++ src/components/Scrollbar.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-vars */
 import React, { useState, useEffect } from 'react';
 
-/** {Component} scrollbar @경로 App/Page/Scrollbar */
+/** Component */
 export default function Scrollbar() {
   return (
     <div className='track'>
