--- src/css/blog.css
@@ -1,1 +1,0 @@
-@charset "utf-8";
