+++ src/styles/body/blog.css
@@ -1,6 +1,4 @@
-/* App.js */
-/* BlogBoby.js */
-/* List.js */
+@import url(blogEdit.css);
 
 .list {
   margin-top: 30px;
