+++ src/styles/body/blogEdit.css
@@ -1,3 +1,5 @@
+/* App/body/blogPosting */
+
 /* 공통 */
 .flexWrap {
   display: flex;
