+++ src/styles/head.css
@@ -1,5 +1,4 @@
-/* App.css */
-/* head.js */
+/* App/head */
 
 header {
   display: flex;
