+++ src/styles/scroll.css
@@ -1,5 +1,6 @@
-@charset "utf-8";
-.page {
+/* App/scroll */
+
+.editor {
   display: flex;
 }
 .viewer {
